Split editor slice selectors
