+++ src/pageEditor/store/editor/editorSelectors/editorClipboardSelectors.ts
@@ -1,0 +1,28 @@
+/*
+ * Copyright (C) 2024 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import type { EditorRootState } from "@/pageEditor/store/editor/pageEditorTypes";
+
+/**
+ * @file Selectors for copy-paste functionality in the editor.
+ */
+
+/**
+ * Select the brick configuration on the clipboard, if any.
+ */
+export const selectCopiedBrick = ({ editor }: EditorRootState) =>
+  editor.copiedBrick;
