+++ src/pageEditor/store/editor/editorSelectors/editorDesignViewSelectors.ts
@@ -1,0 +1,38 @@
+/*
+ * Copyright (C) 2024 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import type { EditorRootState } from "@/pageEditor/store/editor/pageEditorTypes";
+import { DataPanelTabKey } from "@/pageEditor/tabs/editTab/dataPanel/dataPanelTypes";
+import { selectNodeDataPanelTabState } from "@/pageEditor/store/editor/editorSelectors/editorDataPanelSelectors";
+
+/**
+ * @file Selectors for the Form Builder and Document Builder Design view.
+ *
+ * Separate file because the design view will be moved out of the data panel in the future.
+ */
+
+/**
+ * Returns the active element of the Document or Form builder, or null if no element is selected.
+ */
+export function selectActiveBuilderPreviewElement(
+  state: EditorRootState,
+): string | null {
+  return (
+    selectNodeDataPanelTabState(state, DataPanelTabKey.Design)?.activeElement ??
+    null
+  );
+}
