+++ src/pageEditor/store/editor/editorSelectors/editorErrorSelectors.ts
@@ -1,0 +1,27 @@
+/*
+ * Copyright (C) 2024 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import type { EditorRootState } from "@/pageEditor/store/editor/pageEditorTypes";
+import { deserializeError } from "serialize-error";
+
+export const selectErrorState = ({ editor }: EditorRootState) => ({
+  isBetaError: editor.error && editor.beta,
+  editorError: editor.error ? deserializeError(editor.error) : null,
+});
+
+export const selectIsDimensionsWarningDismissed = (state: EditorRootState) =>
+  state.editor.isDimensionsWarningDismissed;
